refactor(frontend): cleanup before merging into main
