+++ frontend/src/components/molecules/editPluginInfo.tsx
@@ -1,6 +1,10 @@
 import { useState } from "react";
-import type { PluginInfoMap } from "../../data/pluginInfoMap.ts";
-import ApplyPluginInfoChangeButton from "../atoms/applyPluginInfoChangeButton.tsx";
+import {
+	type PluginInfoMap,
+	postChangedPluginInfo,
+} from "../../data/pluginInfoMap.ts";
+import InputField from "../atoms/inputField.tsx";
+import SaveChangesButton from "../atoms/saveChangesButton.tsx";
 import type { PluginMap } from "../templates/pluginList.tsx";
 
 export default function EditPluginInfo(props: {
@@ -35,28 +39,64 @@ function NewEditor(
 			<h2 className="text-2xl text-gray-700 mb-2">
 				Edit information of {pluginName}
 			</h2>
-			<h3 className="text-xl text-gray-700 mb-1">Description</h3>
-			<textarea
-				id={`${pluginName}-description`}
-				className="bg-white border border-gray-400 px-1 w-full h-32"
-				onChange={(event) => setDescriptionInput(event.target.value)}
-				defaultValue={infoMap.get("description")?.value ?? ""}
+			<InputField
+				name="Description"
+				input={
+					<textarea
+						id={`${pluginName}-description`}
+						className="bg-white border border-gray-400 px-1 w-full h-32"
+						onChange={(event) => setDescriptionInput(event.target.value)}
+						defaultValue={infoMap.get("description")?.value ?? ""}
+					/>
+				}
 			/>
-			<h3 className="text-xl text-gray-700 mb-1">URL</h3>
-			<input
-				id={`${pluginName}-url`}
-				className="bg-white border border-gray-400 px-1 h-7 w-full"
-				onChange={(event) => setUrlInput(event.target.value)}
-				type="url"
-				defaultValue={infoMap.get("url")?.value ?? ""}
+			<InputField
+				name="URL"
+				input={
+					<input
+						id={`${pluginName}-url`}
+						className="bg-white border border-gray-400 px-1 h-7 w-full"
+						onChange={(event) => setUrlInput(event.target.value)}
+						type="url"
+						defaultValue={infoMap.get("url")?.value ?? ""}
+					/>
+				}
 			/>
-			<ApplyPluginInfoChangeButton
-				pluginName={pluginName}
-				map={infoMap}
-				description={descriptionInput}
-				url={urlInput}
-				refresh={refresh}
+			<SaveChangesButton
+				onClick={() =>
+					applyChanges(pluginName, infoMap, descriptionInput, urlInput, refresh)
+				}
 			/>
 		</div>
 	);
+}
+
+function applyChanges(
+	pluginName: string,
+	map: PluginInfoMap,
+	description: string,
+	url: string,
+	refresh: () => void,
+) {
+	setValue(map, "description", description);
+	setValue(map, "url", url);
+
+	(async () => {
+		await postChangedPluginInfo(pluginName, map);
+	})();
+
+	refresh();
+}
+
+function setValue(map: PluginInfoMap, key: string, value: string) {
+	const holder = map.get(key);
+
+	if (holder === undefined) {
+		map.set(key, {
+			original: "",
+			value: value,
+		});
+	} else {
+		holder.value = value;
+	}
 }
